EP-806. Moved math_utils under computations.
