+++ tvb_epilepsy/service/lsa_service.py
@@ -11,7 +11,7 @@ from tvb_epilepsy.base.utils.log_error_utils import initialize_logger, raise_val
 from tvb_epilepsy.base.utils.data_structures_utils import formal_repr
 from tvb_epilepsy.base.computations.calculations_utils import calc_fz_jac_square_taylor
 from tvb_epilepsy.base.computations.equilibrium_computation import calc_eq_z
-from tvb_epilepsy.base.utils.math_utils import weighted_vector_sum, curve_elbow_point
+from tvb_epilepsy.base.computations.math_utils import weighted_vector_sum, curve_elbow_point
 from tvb_epilepsy.service.hypothesis_builder import HypothesisBuilder
